+++ CHANGELOG.md
@@ -25,6 +25,7 @@ All notable changes to the "prettier-vscode" extension will be documented in thi
 - Registers `.graphql` files as `graphql` language in order to provide formatting. (#989)
 - Ignore files are only read from the workspace root folder to behave the [same as prettier](https://github.com/prettier/prettier/issues/4081).
 - Added configuration option `prettier.prettierPath` to override module resolution.
+- Added configuration option `prettier.configPath` to override configuration file resolution.
 
 ## [2.3.0]